requirements.txt
arq
redis
httpx
requests
google-genai
